Improve objective manager by moving setup into the afterModel hook on the courseobjective router.
