+++ app/components/objective-manager.js
@@ -2,10 +2,9 @@ import Ember from 'ember';
 
 export default Ember.Component.extend({
   classNames: ['objective-manager'],
-  objective: null,
-  objectiveTitle: '',
   parentTitle: '',
   groupTitle: '',
+  title: '',
   objectiveGroups: [],
   selectedGroupId: null,
   multipleGroups: Ember.computed.gt('availableGroups.length', 1),
