+++ app/models/user.js
@@ -52,27 +52,7 @@ var User = DS.Model.extend({
       }
       return first + ' ' + last;
   }.property('firstName', 'lastName'),
-  events: function(){
-      var promises = {
-        offerings: this.get('offerings')
-      };
-      return new Ember.RSVP.hash(promises).then(function(results) {
-        var events = [];
-        results.offerings.forEach(function(offering){
-            events.push(offering);
-        });
-
-        var eventPromises = [];
-        events.forEach(function(event){
-           eventPromises.push(event.get('start'));
-           eventPromises.push(event.get('end'));
-           eventPromises.push(event.get('title'));
-        });
-        return new Ember.RSVP.all(eventPromises).then(function(){
-            return events;
-        });
-      });
-  }.property('offerings.@each', 'offerings.@each.session'),
+  events: [],
   allRelatedCourses: Ember.computed.alias('directedCourses'),
 });
