+++ app/templates/components/objective-manager.hbs
@@ -1,4 +1,4 @@
-<h2> <label>{{objectiveTitle}}:</label> {{{objective.title}}}</h2>
+<h2> <label>{{objectiveTitle}}:</label> {{{title}}}</h2>
 {{#if availableGroups}}
   <h3 class='title'>{{parentTitle}}</h3>
