+++ app/templates/courseobjective.hbs
@@ -3,7 +3,7 @@
   {{course-header course=course editable=false}}
   {{course-overview course=course editable=false}}
   {{objective-manager
-    objective=objective
+    title=model.title
     objectiveGroups=objectiveGroups
     addParent='addParent'
     removeParent='removeParent'
